Makes sure simple_peak return is in integer pixels rather than float pixels

+++ xga/products/phot.py
@@ -1,5 +1,5 @@
 #  This code is a part of XMM: Generate and Analyse (XGA), a module designed for the XMM Cluster Survey (XCS).
-#  Last modified by David J Turner (david.turner@sussex.ac.uk) 14/07/2021, 11:47. Copyright (c) David J Turner
+#  Last modified by David J Turner (david.turner@sussex.ac.uk) 28/07/2021, 12:11. Copyright (c) David J Turner
 
 
 import os
@@ -1039,7 +1039,7 @@ class RateMap(Image):
         if out_unit != pix:
             peak_conv = self.coord_conv(peak_pix, out_unit)
         else:
-            peak_conv = peak_pix
+            peak_conv = peak_pix.astype(int)
 
         # Find if the peak coordinates sit near an edge/chip gap
         edge_flag = self.near_edge(peak_pix)
